+++ fh_fablib/local.py
@@ -2,15 +2,14 @@ from __future__ import unicode_literals
 
 from datetime import datetime
 import os
-import speckenv
 
-from fabric.api import env, execute, get, hosts, settings, task
+from fabric.api import env, execute, hosts, settings, task
 from fabric.colors import green, red
 from fabric.contrib.project import rsync_project
 from fabric.utils import abort, puts
 
 from fh_fablib import confirm, run_local, require_env, require_services
-from fh_fablib.utils import get_random_string, TemporaryDirectory
+from fh_fablib.utils import get_random_string, remote_env
 
 
 @task
@@ -172,12 +171,7 @@ def pull_database():
             ' "%(box_database_local)s" (if it exists)?'):
         return
 
-    mapping = {}
-    with TemporaryDirectory() as d:
-        get('%(box_domain)s/.env' % env, d)
-        speckenv.read_speckenv(os.path.join(d, '.env'), mapping=mapping)
-
-    env.box_remote_db = speckenv.env('DATABASE_URL', mapping=mapping)
+    env.box_remote_db = remote_env('DATABASE_URL')
     if not env.box_remote_db:
         abort(red('Unable to determine the remote DATABASE_URL', bold=True))