+++ CHANGELOG.rst
@@ -11,6 +11,8 @@ Change log
   does) to make it easier to develop inside vagrant boxes with forwarded
   ports.
 - Made ``ALLOWED_HOSTS`` on the server not contain redundant entries.
+- ``server.dump_db`` now uses the connection string from the remote
+  ``.env``
 
 
 0.5 (2017-08-25)